Added an error message for blank label pages.
Basically, I've replaced the 'No posts with label X' message with one of
my own. Had to be replaced because there's no way to actually customize
it. There's no easy indicator for when a label or search is
unsuccessful. For an example see this:
http://blogger2ools.mystady.com/search/label/derp#blog-posts-widget-page-navigation
Here's what I figured out, on error pages the variables 'olderPageUrl'
and 'newerPageUrl' are null. Knowing that made the rest easy to figure
out.